+++ tests/lj/parse-test.cpp
@@ -1,0 +1,61 @@
+#include "gtest/gtest.h"
+
+#include "dislua/dislua.hpp"
+
+namespace lj = dislua::lj;
+
+TEST(LUAJIT_TEST, HEADER_PARSE) {
+  const dislua::uint version = 2;
+  const dislua::uint flags = lj::dump_flags::be;
+  const std::string debug_name = "test";
+
+  lj::parser info;
+  info.version = version;
+  info.header = {
+    .flags = flags,
+    .debug_name = debug_name
+  };
+  info.protos.emplace_back(); // empty prototype
+
+  info.write();
+  ASSERT_NO_THROW(info.read());
+
+  EXPECT_EQ(info.version, version);
+  EXPECT_EQ(info.header.flags, flags);
+  EXPECT_EQ(info.header.debug_name, debug_name);
+}
+
+TEST(LUAJIT_TEST, PROTO_PARSE) {
+  const std::vector<dislua::instruction> ins{
+    { lj::v2::bcops::KSTR, 0, 0 },
+    { lj::v2::bcops::RET0, 0, 0 }
+  };
+  const dislua::proto p = {
+    .flags = lj::proto_flags::varargs,
+    .numparams = 3,
+    .framesize = 1,
+
+    // debug
+    .firstline = 0,
+    .numline = 260,
+
+    .ins = ins,
+    .kgc = { "test" },
+    .knum = { 1.0, 3.125 },
+
+    // debug
+    .lineinfo = { 1, 2 }
+  };
+
+  lj::parser info;
+  info.version = 2;
+  info.header = {
+    .flags = lj::dump_flags::be,
+    .debug_name =  "test"
+  };
+  info.protos.push_back(p);
+
+  info.write();
+  ASSERT_NO_THROW(info.read());
+  ASSERT_EQ(info.protos.front(), p);
+}